+++ react/react/src/cache/Provider.tsx
@@ -1,0 +1,16 @@
+
+import React from 'react'
+import { ProviderCacheContext, useCreateCache } from './ProviderCache'
+
+export const CacheProvider: React.FC<{
+    children: React.ReactNode
+}> = ({ children }) => {
+    const cache = useCreateCache()
+    return (
+        <ProviderCacheContext.Provider value={cache}>
+            {children}
+        </ProviderCacheContext.Provider>
+    )
+}
+
+