+++ react/form/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @einfach/react-form
 
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/react@0.3.22
+  - @einfach/react-utils@0.3.22
+
 ## 0.3.21
 
 ### Patch Changes